Refactor(#32): ChangeFontStyle 컴포넌트 분리
