+++ src/components/editor/BasicEditor.tsx
@@ -1,59 +1,18 @@
 import React, { useRef } from "react";
 import ChangeFont from "./ChangeFont";
 import ChangeFontSize from "./ChangeFontSize";
+import ChangeFontStyle from "./ChangeFontStyle";
 
 const BasicEditor = () => {
   const editorRef = useRef<HTMLDivElement>(null);
-
-  const applyStyle = (style: string): void => {
-    if (editorRef.current) {
-      document.execCommand(style);
-
-      if (editorRef.current) {
-        editorRef.current.focus({ preventScroll: true });
-      }
-    }
-  };
 
   const submitHandler = async () => {
     console.log(editorRef.current?.innerHTML);
   };
 
-  const buttonStyle =
-    "btn ml-2 mr-2 my-5 p-1 border border-solid border-black rounded";
-
-  const initialButtonStyle =
-    "btn ml-5 mr-2 my-5 p-1 border border-solid border-black rounded";
-
   return (
     <>
-      <button className={buttonStyle} onClick={() => applyStyle("bold")}>
-        B
-      </button>
-      <button className={buttonStyle} onClick={() => applyStyle("italic")}>
-        I
-      </button>
-      <button className={buttonStyle} onClick={() => applyStyle("underline")}>
-        U
-      </button>
-      <button
-        className={buttonStyle}
-        onClick={() => applyStyle("strikeThrough")}
-      >
-        S
-      </button>
-      <button
-        className={buttonStyle}
-        onClick={() => applyStyle("insertOrderedList")}
-      >
-        OL
-      </button>
-      <button
-        className={buttonStyle}
-        onClick={() => applyStyle("insertUnorderedList")}
-      >
-        UL
-      </button>
+      <ChangeFontStyle ref={editorRef} />
       <ChangeFont ref={editorRef} />
       <ChangeFontSize ref={editorRef} />
       <div
@@ -62,9 +21,12 @@ const BasicEditor = () => {
         contentEditable
         className={`mx-5 p-10 border border-black border-solid rounded min-h-32 [&>img]:w-full`}
       ></div>
-      <button className={initialButtonStyle} onClick={submitHandler}>
+      <button
+        className="btn ml-5 mr-2 my-5 p-1 border border-solid border-black rounded"
+        onClick={submitHandler}
+      >
         submit
-      </button>{" "}
+      </button>
     </>
   );
 };
